Reorganizaçao de pastas e componentes
