+++ src/components/DataDisplay/Avatar/style.css
