+++ src/components/DataDisplay/Tables/Clients/clients-paying/index.jsx
